Add files via upload
Data Scientists at Big Company Inc are rated as either performing well or poorly. This is the performance
variable in the data 1 means performing well and 0 means performing poorly. They have three measures
on them at the time of hiring. Credentials, Age and Test Scores. Credentials are related to their academic
qualifications via a complex system of rating various degree programs. Age is their age in years and Test
Scores is the value of a competency test taken when starting.
Your task is to fit a logistic regression model to predict performance from the other variables. This model
will be used in future hiring decisions!
1. Show your data exploration step and show your rea